added script for ubuntu 24.04 odoo 18.0

+++ odoo18_ubuntu_24.04.sh
@@ -402,8 +402,11 @@ EOF
             echo "Odoo Nginx configuration file is at /etc/nginx/conf.d/odoo.conf ..."
         fi
 
+        if [ -f "/etc/nginx/conf.d/default.conf" ]; then
+            sudo rm /etc/nginx/conf.d/default.conf
+        fi
+
         sudo mv ~/odoo.conf /etc/nginx/conf.d/
-        sudo rm /etc/nginx/conf.d/default.conf
         sudo service nginx restart
         sudo su root -c "printf 'proxy_mode = True\n' >> /etc/${OE_CONFIG}.conf"
         log "INFO" "Nginx server is up and running. Configuration can be found at /etc/nginx/conf.d/odoo.conf"
